# docs/api/openapi3.yaml
openapi: 3.0.1
info:
  title: sync
  description: ""
  version: 0.0.1
servers:
- url: ""
...
paths:
...
  /posts:
    get:
      tags:
      - post
      summary: Get Posts
      description: Get Posts
      operationId: GetPosts
      parameters:
      - name: first
        in: query
        description: Number of items to return from the beginning of the list
        required: false
        schema:
          type: string
      - name: after
        in: query
        description: "Cursor for pagination, indicating the position after which to\
          \ return items"
        required: false
        schema:
          type: string
      - name: last
        in: query
        description: Number of items to return from the end of the list
        required: false
        schema:
          type: string
      - name: before
        in: query
        description: "Cursor for pagination, indicating the position before which\
          \ to return items"
        required: false
        schema:
          type: string
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedGetPostsResponse"
              examples:
                GetPosts:
                  value: "{\"posts\":{\"pageInfo\":{\"size\":1,\"hasNextPage\":false,\"\
                    hasPreviousPage\":false,\"startCursor\":\"start-cursor\",\"endCursor\"\
                    :\"end-cursor\"},\"nodes\":[{\"cursor\":\"cursor\",\"content\"\
                    :{\"id\":1,\"slug\":\"test-slug\",\"title\":\"Test Post Title\"\
                    ,\"type\":\"SHORT\",\"status\":\"PUBLISHED\",\"scope\":\"WORKSPACE\"\
                    ,\"accessLevel\":\"FULL\",\"author\":{\"handle\":\"john-doe\"\
                    ,\"name\":\"John Doe\",\"profileImageUrl\":\"https://example.com/john.png\"\
                    },\"project\":{\"handle\":\"my-project\",\"name\":\"나의 프로젝트\"\
                    ,\"description\":\"프로젝트 설명\",\"website\":\"https://example.com\"\
                    ,\"isPublic\":true,\"joinPolicy\":\"INVITE\",\"followerCount\"\
                    :42,\"iconUrl\":\"https://example.com/icon.png\"},\"resolved\"\
                    :false,\"isSeriesPost\":false,\"pinnedAt\":null,\"isAuthor\":false,\"\
                    canDelete\":false,\"canComment\":true,\"createdAt\":\"1970-01-01T00:00:00Z\"\
                    ,\"updatedAt\":\"1970-01-01T00:00:00Z\",\"likeCount\":1,\"liked\"\
                    :true,\"commentCount\":1,\"bookmarked\":true,\"tags\":[{\"id\"\
                    :1,\"name\":\"java\",\"description\":\"자바 관련 태그\",\"postCount\"\
                    :10,\"followerCount\":3,\"projectHandle\":null,\"isFollowing\"\
                    :false}],\"preview\":\"This is a preview of the post content.\"\
                    ,\"previewMedia\":[{\"id\":1,\"url\":\"https://example.com/image.png\"\
                    ,\"fileName\":\"image.png\",\"fileSize\":102400,\"mediaType\"\
                    :\"image/png\"}],\"mediaCount\":1,\"wordCount\":120,\"coverImageUrl\"\
                    :\"https://example.com/cover.png\"}}]}}"
    post:
      tags:
      - post
      summary: Create Post
      description: Create Post
      operationId: CreatePost
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePostRequest"
            examples:
              CreatePost:
                value: "{\"title\":\"title\",\"type\":\"SHORT\",\"status\":\"PUBLISHED\"\
                  ,\"content\":{\"text\":\"This is a post content.\",\"json\":\"{\\\
                  \"text\\\": \\\"This is a post content.\\\"}\",\"mediaIds\":[1,2]},\"\
                  tags\":[\"java\",\"spring\"],\"referencedPostIds\":[10,11],\"coverMediaId\"\
                  :\"100\"}"
      responses:
        "201":
          description: "201"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatePostResponse"
              examples:
                CreatePost:
                  value: |-
                    {
                      "slug" : "slug"
                    }
...
  /posts/{postId}:
    delete:
      tags:
      - post
      summary: Delete Post
      description: Delete Post
      operationId: DeletePost
      parameters:
      - name: postId
        in: path
        description: Post ID
        required: true
        schema:
          type: string
      responses:
        "204":
          description: "204"
    patch:
      tags:
      - post
      summary: Update Post
      description: Update Post
      operationId: UpdatePost
      parameters:
      - name: postId
        in: path
        description: Post ID
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePostRequest"
            examples:
              UpdatePost:
                value: "{\"title\":\"Updated post title\",\"type\":\"LONG\",\"status\"\
                  :\"PUBLISHED\",\"content\":{\"text\":\"This is a post content.\"\
                  ,\"json\":\"{\\\"text\\\": \\\"This is a post content.\\\"}\",\"\
                  mediaIds\":[1]},\"tags\":[\"java\"],\"referencedPostIds\":[10,11],\"\
                  coverMediaId\":\"100\",\"removeCover\":false}"
      responses:
        "204":
          description: "204"
  /posts/{slug}:
    get:
      tags:
      - post
      summary: Get Post By Slug
      description: Get Post By Slug
      operationId: GetPostBySlug
      parameters:
      - name: slug
        in: path
        description: ""
        required: true
        schema:
          type: string
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetPostResponse"
              examples:
                GetPostBySlug:
                  value: "{\"summary\":{\"id\":1,\"slug\":\"test-slug\",\"title\"\
                    :\"Test Post Title\",\"type\":\"SHORT\",\"status\":\"PUBLISHED\"\
                    ,\"scope\":\"WORKSPACE\",\"accessLevel\":\"FULL\",\"author\":{\"\
                    handle\":\"john-doe\",\"name\":\"John Doe\",\"profileImageUrl\"\
                    :\"https://example.com/john.png\"},\"project\":{\"handle\":\"\
                    my-project\",\"name\":\"나의 프로젝트\",\"description\":\"프로젝트 설명\"\
                    ,\"website\":\"https://example.com\",\"isPublic\":true,\"joinPolicy\"\
                    :\"INVITE\",\"followerCount\":42,\"iconUrl\":\"https://example.com/icon.png\"\
                    },\"resolved\":false,\"isSeriesPost\":false,\"pinnedAt\":null,\"\
                    isAuthor\":false,\"canDelete\":false,\"canComment\":true,\"createdAt\"\
                    :\"1970-01-01T00:00:00Z\",\"updatedAt\":\"1970-01-01T00:00:00Z\"\
                    ,\"likeCount\":1,\"liked\":true,\"commentCount\":1,\"bookmarked\"\
                    :true,\"tags\":[{\"id\":1,\"name\":\"java\",\"description\":\"\
                    자바 관련 태그\",\"postCount\":10,\"followerCount\":3,\"projectHandle\"\
                    :null,\"isFollowing\":false}],\"preview\":\"This is a preview\
                    \ of the post content.\",\"previewMedia\":[{\"id\":1,\"url\":\"\
                    https://example.com/image.png\",\"fileName\":\"image.png\",\"\
                    fileSize\":102400,\"mediaType\":\"image/png\"}],\"mediaCount\"\
                    :1,\"wordCount\":120,\"coverImageUrl\":\"https://example.com/cover.png\"\
                    },\"content\":{\"json\":\"Post Content\",\"media\":[{\"id\":1,\"\
                    url\":\"https://example.com/media.png\",\"fileName\":\"media.png\"\
                    ,\"fileSize\":102400,\"mediaType\":\"image/png\"}]}}"
...
  /projects/{handle}/posts:
    get:
      tags:
      - post
      summary: Get Posts By Project
      description: Get Posts By Project
      operationId: GetPostsByProject
      parameters:
      - name: handle
        in: path
        description: 프로젝트 핸들
        required: true
        schema:
          type: string
      - name: first
        in: query
        description: Number of items to return from the beginning of the list
        required: false
        schema:
          type: string
      - name: after
        in: query
        description: "Cursor for pagination, indicating the position after which to\
          \ return items"
        required: false
        schema:
          type: string
      - name: last
        in: query
        description: Number of items to return from the end of the list
        required: false
        schema:
          type: string
      - name: before
        in: query
        description: "Cursor for pagination, indicating the position before which\
          \ to return items"
        required: false
        schema:
          type: string
      - name: type
        in: query
        description: 게시글 타입
        required: false
        schema:
          type: string
      - name: authorHandle
        in: query
        description: 작성자 핸들
        required: false
        schema:
          type: string
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedGetPostsResponse"
              examples:
                GetPostsByProject:
                  value: "{\"posts\":{\"pageInfo\":{\"size\":1,\"hasNextPage\":false,\"\
                    hasPreviousPage\":false,\"startCursor\":\"start-cursor\",\"endCursor\"\
                    :\"end-cursor\"},\"nodes\":[{\"cursor\":\"cursor\",\"content\"\
                    :{\"id\":1,\"slug\":\"test-slug\",\"title\":\"Test Post Title\"\
                    ,\"type\":\"SHORT\",\"status\":\"PUBLISHED\",\"scope\":\"WORKSPACE\"\
                    ,\"accessLevel\":\"FULL\",\"author\":{\"handle\":\"john-doe\"\
                    ,\"name\":\"John Doe\",\"profileImageUrl\":\"https://example.com/john.png\"\
                    },\"project\":{\"handle\":\"my-project\",\"name\":\"나의 프로젝트\"\
                    ,\"description\":\"프로젝트 설명\",\"website\":\"https://example.com\"\
                    ,\"isPublic\":true,\"joinPolicy\":\"INVITE\",\"followerCount\"\
                    :42,\"iconUrl\":\"https://example.com/icon.png\"},\"resolved\"\
                    :false,\"isSeriesPost\":false,\"pinnedAt\":null,\"isAuthor\":false,\"\
                    canDelete\":false,\"canComment\":true,\"createdAt\":\"1970-01-01T00:00:00Z\"\
                    ,\"updatedAt\":\"1970-01-01T00:00:00Z\",\"likeCount\":1,\"liked\"\
                    :true,\"commentCount\":1,\"bookmarked\":true,\"tags\":[{\"id\"\
                    :1,\"name\":\"java\",\"description\":\"자바 관련 태그\",\"postCount\"\
                    :10,\"followerCount\":3,\"projectHandle\":null,\"isFollowing\"\
                    :false}],\"preview\":\"This is a preview of the post content.\"\
                    ,\"previewMedia\":[{\"id\":1,\"url\":\"https://example.com/image.png\"\
                    ,\"fileName\":\"image.png\",\"fileSize\":102400,\"mediaType\"\
                    :\"image/png\"}],\"mediaCount\":1,\"wordCount\":120,\"coverImageUrl\"\
                    :\"https://example.com/cover.png\"}}]}}"
    post:
      tags:
      - post
      summary: Create Project Post
      description: 프로젝트에 글을 작성합니다.
      operationId: CreateProjectPost
      parameters:
      - name: handle
        in: path
        description: 프로젝트 핸들
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectPostRequest"
            examples:
              CreateProjectPost:
                value: "{\"title\":\"title\",\"type\":\"SHORT\",\"status\":\"PUBLISHED\"\
                  ,\"content\":{\"text\":\"This is a post content.\",\"json\":\"{\\\
                  \"text\\\": \\\"This is a post content.\\\"}\",\"mediaIds\":[1,2]},\"\
                  tags\":[\"java\",\"spring\"],\"projectTags\":[\"roadmap\"],\"referencedPostIds\"\
                  :[10,11],\"coverMediaId\":\"100\"}"
      responses:
        "201":
          description: "201"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatePostResponse"
              examples:
                CreateProjectPost:
                  value: |-
                    {
                      "slug" : "slug"
                    }
...
  /projects/{handle}/posts/{postId}:
    patch:
      tags:
      - post
      summary: Update Project Post
      description: 프로젝트 글을 수정합니다.
      operationId: UpdateProjectPost
      parameters:
      - name: handle
        in: path
        description: 프로젝트 핸들
        required: true
        schema:
          type: string
      - name: postId
        in: path
        description: Post ID
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectPostRequest"
            examples:
              UpdateProjectPost:
                value: "{\"title\":\"Updated post title\",\"type\":\"LONG\",\"status\"\
                  :\"PUBLISHED\",\"content\":{\"text\":\"This is a post content.\"\
                  ,\"json\":\"{\\\"text\\\": \\\"This is a post content.\\\"}\",\"\
                  mediaIds\":[1]},\"tags\":[\"java\"],\"projectTags\":[\"roadmap\"\
                  ],\"referencedPostIds\":[10,11],\"coverMediaId\":\"100\",\"removeCover\"\
                  :false}"
      responses:
        "204":
          description: "204"
...
  /projects/{handle}/posts/{postId}/pin:
    post:
      tags:
      - post
      summary: Pin Post
      description: 프로젝트 관리자가 게시글을 프로젝트 대시보드에 고정합니다.
      operationId: PinPost
      parameters:
      - name: handle
        in: path
        description: 프로젝트 핸들
        required: true
        schema:
          type: string
      - name: postId
        in: path
        description: Post ID
        required: true
        schema:
          type: string
      responses:
        "204":
          description: "204"
    delete:
      tags:
      - post
      summary: Unpin Post
      description: 프로젝트 관리자가 게시글의 고정을 해제합니다.
      operationId: UnpinPost
      parameters:
      - name: handle
        in: path
        description: 프로젝트 핸들
        required: true
        schema:
          type: string
      - name: postId
        in: path
        description: Post ID
        required: true
        schema:
          type: string
      responses:
        "204":
          description: "204"
...
components:
  schemas:
...
    CreatePostResponse:
      title: CreatePostResponse
      required:
      - slug
      type: object
      properties:
        slug:
          type: string
          description: Post Slug
...
    PaginatedGetPostsResponse:
      title: PaginatedGetPostsResponse
      type: object
      properties:
        posts:
          required:
          - nodes
          - pageInfo
          type: object
          properties:
            nodes:
              type: array
              description: Nodes
              items:
                required:
                - content
                - cursor
                type: object
                properties:
                  cursor:
                    type: string
                    description: Node Cursor
                  content:
                    required:
                    - accessLevel
                    - author
                    - bookmarked
                    - canComment
                    - canDelete
                    - commentCount
                    - createdAt
                    - id
                    - isAuthor
                    - isSeriesPost
                    - likeCount
                    - liked
                    - mediaCount
                    - preview
                    - previewMedia
                    - resolved
                    - scope
                    - slug
                    - status
                    - tags
                    - type
                    - updatedAt
                    - wordCount
                    type: object
                    properties:
                      preview:
                        type: string
                        description: 게시물 내용의 일반 텍스트 미리보기
                      coverImageUrl:
                        type: string
                        description: 게시물 커버 이미지 URL (없으면 없음)
                        nullable: true
                      project:
                        type: object
                        properties:
                          website:
                            type: string
                            description: 프로젝트 웹사이트
                            nullable: true
                          name:
                            type: string
                            description: 프로젝트 이름
                            nullable: true
                          isPublic:
                            type: boolean
                            description: 공개 여부
                            nullable: true
                          description:
                            type: string
                            description: 프로젝트 설명
                            nullable: true
                          handle:
                            type: string
                            description: 프로젝트 핸들
                            nullable: true
                          iconUrl:
                            type: string
                            description: 프로젝트 아이콘 URL
                            nullable: true
                          joinPolicy:
                            type: string
                            description: 프로젝트 참여 정책
                            nullable: true
                            enum:
                            - OPEN
                            - REQUEST
                            - INVITE
                          followerCount:
                            type: number
                            description: 프로젝트 팔로워 수
                            nullable: true
                        description: 소속 프로젝트 정보
                      likeCount:
                        type: number
                        description: Number of Likes
                      type:
                        type: string
                        description: Post Type
                        enum:
                        - SHORT
                        - LONG
                        - QUESTION
                      title:
                        type: string
                        description: Post Title
                        nullable: true
                      liked:
                        type: boolean
                        description: Whether the current user liked this post
                      isAuthor:
                        type: boolean
                        description: Whether the requesting user is the author of
                          this post
                      pinnedAt:
                        type: string
                        description: 게시글이 프로젝트 대시보드에 고정된 시각 (고정되지 않은 경우 없음)
                        nullable: true
                      createdAt:
                        type: string
                        description: Creation Timestamp
                      scope:
                        type: string
                        description: 게시글 공개 범위
                        enum:
                        - PUBLIC
                        - WORKSPACE
                      mediaCount:
                        type: number
                        description: 게시물에 첨부된 전체 미디어 수
                      canDelete:
                        type: boolean
                        description: "요청자가 이 게시글을 삭제할 수 있는지 여부 (작성자, 플랫폼 관리자(개인 게시\
                          글), 프로젝트 관리자(프로젝트 게시글))"
                      canComment:
                        type: boolean
                        description: 요청자가 이 게시글에 댓글을 작성할 수 있는지 여부 (프로젝트 게시글은 팀원만 가능)
                      id:
                        type: number
                        description: Post ID
                      slug:
                        type: string
                        description: Post Slug
                      updatedAt:
                        type: string
                        description: Last Updated Timestamp
                      resolved:
                        type: boolean
                        description: Whether the question post has been resolved
                      wordCount:
                        type: number
                        description: 게시물 본문의 단어 수
                      accessLevel:
                        type: string
                        description: "요청자의 열람 수준 (FULL: 본문까지 열람, PREVIEW: 유료 게이트로\
                          \ 본문 잠김)"
                        enum:
                        - FULL
                        - PREVIEW
                      bookmarked:
                        type: boolean
                        description: Whether the current user bookmarked this post
                      author:
                        required:
                        - handle
                        - name
                        type: object
                        properties:
                          name:
                            type: string
                            description: 유저 이름
                          handle:
                            type: string
                            description: 유저 핸들
                          profileImageUrl:
                            type: string
                            description: 유저 프로필 이미지 URL
                            nullable: true
                        description: 작성자 정보
                      tags:
                        type: array
                        description: 게시물에 달린 태그 목록
                        items:
                          required:
                          - description
                          - followerCount
                          - id
                          - isFollowing
                          - name
                          - postCount
                          type: object
                          properties:
                            isFollowing:
                              type: boolean
                              description: 요청자가 해당 태그를 팔로우하고 있는지 여부 (프로젝트 태그는 항상 false)
                            name:
                              type: string
                              description: 태그 이름
                            postCount:
                              type: number
                              description: 태그가 사용된 게시물 수
                            description:
                              type: string
                              description: 태그 설명
                            id:
                              type: number
                              description: 태그 ID
                            followerCount:
                              type: number
                              description: 태그를 팔로우하는 사용자 수
                            projectHandle:
                              type: string
                              description: 프로젝트 태그인 경우 해당 프로젝트의 핸들 (전역 태그인 경우 없음)
                              nullable: true
                      commentCount:
                        type: number
                        description: Number of Comments
                      isSeriesPost:
                        type: boolean
                        description: 게시글이 어떤 시리즈에 속해 있는지 여부
                      previewMedia:
                        type: array
                        description: 미리보기용 첨부 이미지 목록 (최대 3개). 이미지가 아닌 첨부 파일은 포함되지
                          않는다
                        items:
                          required:
                          - fileName
                          - fileSize
                          - id
                          - mediaType
                          - url
                          type: object
                          properties:
                            fileName:
                              type: string
                              description: 업로드된 원본 파일 이름
                            fileSize:
                              type: number
                              description: 파일 크기 (바이트)
                            mediaType:
                              type: string
                              description: 파일의 MIME 타입
                            id:
                              type: number
                              description: 미디어 ID
                            url:
                              type: string
                              description: 미디어 URL
                      status:
                        type: string
                        description: Post Status
                        enum:
                        - DRAFT
                        - PUBLISHED
                    description: Node Content
            pageInfo:
              required:
              - hasNextPage
              - hasPreviousPage
              - size
              type: object
              properties:
                size:
                  type: number
                  description: Page Size
                hasNextPage:
                  type: boolean
                  description: Has Next Page
                hasPreviousPage:
                  type: boolean
                  description: Has Previous Page
                endCursor:
                  type: string
                  description: End Cursor
                  nullable: true
                startCursor:
                  type: string
                  description: Start Cursor
                  nullable: true
              description: Page Info
...
    CreatePostRequest:
      title: CreatePostRequest
      required:
      - content
      - type
      type: object
      properties:
        coverMediaId:
          type: string
          description: 커버 이미지로 사용할 미디어 ID (없으면 커버 없음)
          nullable: true
        referencedPostIds:
          type: array
          description: "이 글이 참조하는 게시글 ID 목록 (지정한 순서대로 노출, 최대 50개)"
          nullable: true
          items:
            oneOf:
            - type: object
            - type: boolean
            - type: string
            - type: number
        type:
          type: string
          description: Post Type
          enum:
          - SHORT
          - LONG
          - QUESTION
        title:
          type: string
          description: Title
          nullable: true
        content:
          required:
          - json
          - text
          type: object
          properties:
            json:
              type: string
              description: JSON Content
            mediaIds:
              type: array
              description: 사용된 미디어 ID 목록
              nullable: true
              items:
                oneOf:
                - type: object
                - type: boolean
                - type: string
                - type: number
            text:
              type: string
              description: Text Content
          description: Content
        tags:
          type: array
          description: 태그 목록
          nullable: true
          items:
            oneOf:
            - type: object
            - type: boolean
            - type: string
            - type: number
        status:
          type: string
          description: Post Status
          nullable: true
          enum:
          - DRAFT
          - PUBLISHED
...
    UpdatePostRequest:
      title: UpdatePostRequest
      required:
      - content
      - status
      - type
      type: object
      properties:
        coverMediaId:
          type: string
          description: 새 커버 이미지로 사용할 미디어 ID (null 이면 기존 커버 유지)
          nullable: true
        removeCover:
          type: boolean
          description: true 이면 기존 커버 이미지를 제거한다
          nullable: true
        referencedPostIds:
          type: array
          description: "이 글이 참조하는 게시글 ID 목록 (지정한 순서대로 노출, 최대 50개)"
          nullable: true
          items:
            oneOf:
            - type: object
            - type: boolean
            - type: string
            - type: number
        type:
          type: string
          description: Post Type
          enum:
          - SHORT
          - LONG
          - QUESTION
        title:
          type: string
          description: Title
          nullable: true
        content:
          required:
          - json
          - text
          type: object
          properties:
            json:
              type: string
              description: JSON Content
            mediaIds:
              type: array
              description: 사용된 미디어 ID 목록
              nullable: true
              items:
                oneOf:
                - type: object
                - type: boolean
                - type: string
                - type: number
            text:
              type: string
              description: Text Content
          description: Content
        tags:
          type: array
          description: 전역 태그 목록
          nullable: true
          items:
            oneOf:
            - type: object
            - type: boolean
            - type: string
            - type: number
        status:
          type: string
          description: Post Status
          enum:
          - DRAFT
          - PUBLISHED
...
    CreateProjectPostRequest:
      title: CreateProjectPostRequest
      required:
      - content
      - type
      type: object
      properties:
        coverMediaId:
          type: string
          description: 커버 이미지로 사용할 미디어 ID (없으면 커버 없음)
          nullable: true
        projectTags:
          type: array
          description: 프로젝트 태그 목록
          nullable: true
          items:
            oneOf:
            - type: object
            - type: boolean
            - type: string
            - type: number
        referencedPostIds:
          type: array
          description: "이 글이 참조하는 게시글 ID 목록 (지정한 순서대로 노출, 최대 50개)"
          nullable: true
          items:
            oneOf:
            - type: object
            - type: boolean
            - type: string
            - type: number
        type:
          type: string
          description: Post Type
          enum:
          - SHORT
          - LONG
          - QUESTION
        title:
          type: string
          description: Title
          nullable: true
        content:
          required:
          - json
          - text
          type: object
          properties:
            json:
              type: string
              description: JSON Content
            mediaIds:
              type: array
              description: 사용된 미디어 ID 목록
              nullable: true
              items:
                oneOf:
                - type: object
                - type: boolean
                - type: string
                - type: number
            text:
              type: string
              description: Text Content
          description: Content
        tags:
          type: array
          description: 전역 태그 목록
          nullable: true
          items:
            oneOf:
            - type: object
            - type: boolean
            - type: string
            - type: number
        status:
          type: string
          description: Post Status
          nullable: true
          enum:
          - DRAFT
          - PUBLISHED
...